In this line, one should summarize the changes
And in this paragraph, one can go into detail about the changes made. Keep it wrapped to around 72 characters. Make sure to have a blank line between the first summarizing line and the explanatory body text.

The point of these body paragraphs are to explain why you are making this change. For example, if one is adding alt text to an image, here they could write, "screen readers won't read the images to users with disabilities without this information" as an explanation.

You can add more paragraphs, but ensure there is a blank line between each. You can also use bullet points:

 - A hyphen or asterisk is used for the bullet.

 - Make sure to include a space before and after the bullet point.

 - Also include a blank line between each point, like with the paragraphs.

If you are using an issue tracker, you can include a reference to it at the bottom like so:

Resolves: #123
See also: #456, #789